requirements.txt
Flask==3.1.0
google-cloud-storage==2.19.0
google-generativeai==0.8.4
gunicorn==23.0.0